% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastLasso.R
\name{fastLasso}
\alias{fastLasso}
\title{Lasso regression on a subset}
\usage{
fastLasso(
  x,
  y,
  lambda,
  subset = NULL,
  normalize = TRUE,
  intercept = TRUE,
  eps = .Machine$double.eps,
  use.Gram = TRUE,
  drop = TRUE,
  raw = FALSE
)
}
\arguments{
\item{x}{a numeric matrix containing the predictor variables.}

\item{y}{a numeric vector containing the response variable.}

\item{lambda}{a non-negative numeric value to be used as penalty parameter.}

\item{subset}{an integer vector giving the subset on which to compute the
lasso.  If \code{NULL}, the lasso is computed on the full data set.}

\item{normalize}{a logical indicating whether the predictor variables
should be normalized to have unit \eqn{L_{2}}{L2} norm (the default is
\code{TRUE}).  Note that normalization is performed on the subsample
rather than the full data set.}

\item{intercept}{a logical indicating whether a constant term should be
included in the model (the default is \code{TRUE}).}

\item{eps}{a small positive numeric value used to determine whether the
variability within a variable is too small (an effective zero).}

\item{use.Gram}{a logical indicating whether the Gram matrix of the
explanatory variables should be precomputed on the subsample (the default is
\code{TRUE}).  If the number of variables is large, computation may be
faster when this is set to \code{FALSE}.}

\item{drop}{a logical indicating whether relevant components of the output
should be returned as vectors (\code{TRUE}, the default) or matrices
(\code{FALSE}).}

\item{raw}{argument for internal use that is not expected to be used by
others. (If \code{TRUE}, somewhat different output is returned that is
relevant for the raw fit without trimming in \code{\link{sparseLTS}()}.)}
}
\value{
A list with the following components:
\describe{
  \item{\code{coefficients}}{a numeric vector or matrix containing the
  coefficient estimates (as determined by argument \code{drop}).}
  \item{\code{fitted.values}}{a numeric vector or matrix containing the
  fitted values (as determined by argument \code{drop}).}
  \item{\code{residuals}}{a numeric vector or matrix containing the
  residuals (as determined by argument \code{drop}).}
}
}
\description{
Wrapper function for a barebones implementation in \proglang{C++} to
compute the lasso on a subset of the data for a fixed value of the penalty
parameter via the LARS algorithm. It is mainly used internally by
\code{\link{sparseLTS}()} to compute the reweighted fit, but also for the
raw fit in the rare case of no trimming.
}
\note{
As of version 0.8.5, this function is exported for use by other developers
due to popular demand. Use this function only if you know what you are
doing! There is no error handling, and the underlying \proglang{C++} code
may crash the \proglang{R} session if improper arguments are supplied.
}
\references{
Efron, B., Hastie, T., Johnstone, I. and Tibshirani, R. (2003) Least angle
regression. \emph{The Annals of Statistics}, \bold{32}(2), 407--499.
\doi{10.1214/009053604000000067}
}
\seealso{
\code{\link{sparseLTS}}
}
\author{
Andreas Alfons
}
\keyword{regression}
