% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustHD-package.R
\docType{package}
\name{robustHD-package}
\alias{robustHD}
\alias{robustHD-package}
\title{\packageTitle{robustHD}}
\description{
\packageDescription{robustHD}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{robustHD}
\packageIndices{robustHD}
}
\references{
Alfons (2021) \pkg{robustHD}: An \R package for robust regression with
high-dimensional data. \emph{Journal of Open Source Software}, \bold{6}(67),
3786. \doi{10.21105/joss.03786}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aalfons/robustHD}
  \item Report bugs at \url{https://github.com/aalfons/robustHD/issues}
}

}
\author{
\packageAuthor{robustHD}

Maintainer: \packageMaintainer{robustHD}
}
\keyword{package}
