\name{rgrplars}
\alias{rgrplars}
\alias{rgrplars.data.frame}
\alias{rgrplars.default}
\alias{rgrplars.formula}
\title{Robust groupwise least angle regression}
\usage{
  rgrplars(x, ...)

  \method{rgrplars}{formula} (formula, data, ...)

  \method{rgrplars}{data.frame} (x, y, ...)

  \method{rgrplars}{default} (x, y, sMax = NA, assign,
    dummy, centerFun = median, scaleFun = mad,
    regFun = lmrob, regArgs = list(),
    combine = c("min", "euclidean", "mahalanobis"),
    const = 2, prob = 0.95, fit = TRUE, crit = "BIC",
    ncores = 1, cl = NULL, model = TRUE, ...)
}
\arguments{
  \item{formula}{a formula describing the full model.}

  \item{data}{an optional data frame, list or environment
  (or object coercible to a data frame by
  \code{\link{as.data.frame}}) containing the variables in
  the model.  If not found in data, the variables are taken
  from \code{environment(formula)}, typically the
  environment from which \code{rgrplars} is called.}

  \item{x}{a matrix or data frame containing the candidate
  predictors.}

  \item{y}{a numeric vector containing the response.}

  \item{sMax}{an integer vector of length two.  If a single
  integer is supplied, it is recycled.  The first element
  gives the number of predictor groups to be sequenced.  If
  it is \code{NA} (the default), predictor groups are
  sequenced as long as there are no singularity issues.
  The second element gives the maximum number of predictor
  groups to be included in the final model.  If it is
  \code{NA} (the default), predictor groups may be added to
  the model as long as there are twice as many observations
  as predictor variables.}

  \item{assign}{an integer vector giving the predictor
  group to which each predictor variable belongs.}

  \item{dummy}{a logical vector indicating whether the
  predictors are dummy variables.}

  \item{centerFun}{a function to compute a robust estimate
  for the center (defaults to
  \code{\link[stats]{median}}).}

  \item{scaleFun}{a function to compute a robust estimate
  for the scale (defaults to \code{\link[stats]{mad}}).}

  \item{regFun}{a function to compute robust linear
  regressions that can be interpreted as weighted least
  squares (defaults to \code{\link[robustbase]{lmrob}}).}

  \item{regArgs}{a list of arguments to be passed to
  \code{regFun}.}

  \item{combine}{a character string specifying how to
  combine the data cleaning weights from the robust
  regressions with each predictor group. Possible values
  are \code{"min"} for taking the minimum weight for each
  observation, \code{"euclidean"} for weights based on
  Euclidean distances of the multivariate set of
  standardized residuals (i.e., multivariate winsorization
  of the standardized residuals assuming independence), or
  \code{"mahalanobis"} for weights based on Mahalanobis
  distances of the multivariate set of standardized
  residuals (i.e., multivariate winsorization of the
  standardized residuals).}

  \item{const}{numeric; tuning constant for multivariate
  winsorization to be used in the initial corralation
  estimates based on adjusted univariate winsorization
  (defaults to 2).}

  \item{prob}{numeric; probability for the quantile of the
  \eqn{\chi^{2}}{chi-squared} distribution to be used in
  multivariate winsorization (defaults to 0.95).}

  \item{fit}{a logical indicating whether to fit submodels
  along the sequence (\code{TRUE}, the default) or to
  simply return the sequence (\code{FALSE}).}

  \item{crit}{a character string specifying the optimality
  criterion to be used for selecting the final model.
  Currently, only \code{"BIC"} for the Bayes information
  criterion is implemented.}

  \item{ncores}{a positive integer giving the number of
  processor cores to be used for parallel computing (the
  default is 1 for no parallelization).  If this is set to
  \code{NA}, all available processor cores are used.}

  \item{cl}{a \pkg{parallel} cluster for parallel computing
  as generated by \code{\link[parallel]{makeCluster}}.
  This is ignored if \code{ncores} is supplied.}

  \item{model}{a logical indicating whether the model data
  should be included in the returned object.}

  \item{\dots}{additional arguments to be passed down.}
}
\value{
  If \code{fit} is \code{FALSE}, an integer vector
  containing the indices of the sequenced predictor groups.

  Otherwise an object of class \code{"grplars"} (inheriting
  from class \code{"seqModel"}) with the following
  components:

  \item{active}{an integer vector containing the sequence
  of predictor groups.}

  \item{coefficients}{a numeric matrix in which each column
  contains the regression coefficients of the corresponding
  submodel along the sequence.}

  \item{fitted.values}{a numeric matrix in which each
  column contains the fitted values of the corresponding
  submodel along the sequence.}

  \item{residuals}{a numeric matrix in which each column
  contains the residuals of the corresponding submodel
  along the sequence.}

  \item{crit}{a character string specifying the optimality
  criterion used for selecting the final model.}

  \item{critValues}{a numeric vector containing the values
  of the optimality criterion from the submodels along the
  sequence.}

  \item{df}{an integer vector containing the degrees of
  freedom of the submodels along the sequence (i.e., the
  number of estimated coefficients).}

  \item{sOpt}{an integer giving the optimal submodel.}

  \item{muY}{numeric; the mean of the cleaned response.}

  \item{sigmaY}{numeric; the standard deviation of the
  cleaned response.}

  \item{muX}{a numeric vector containing the means of the
  cleaned predictor variables.}

  \item{sigmaX}{a numeric vector containing the standard
  deviations of the cleaned predictor variables.}

  \item{x}{the matrix of candidate predictors (if
  \code{model} is \code{TRUE}).}

  \item{y}{the response (if \code{model} is \code{TRUE}).}

  \item{assign}{an integer vector giving the predictor
  group to which each predictor variable belongs.}

  \item{robust}{a logical indicating whether a robust fit
  was computed (\code{TRUE} for function \code{rgrplars}).}

  \item{w}{a numeric vector giving the data cleaning
  weights.}

  \item{call}{the matched function call.}
}
\description{
  Robustly sequence groups of candidate predictors
  according to their predictive content and find the
  optimal model along the sequence.
}
\note{
  Parallel computing for obtaining the initial data
  cleaning weights is implemented on the \R level via
  package \pkg{parallel}.  Since common robust regression
  function such as \code{\link[robustbase]{lmrob}} (the
  default) involve some random component, it is recommended
  to set the random number generator to
  \code{RNGkind("L'Ecuyer-CMRG")} beforehand for
  reprodicibility.

  In addition, parallel computing for some of of the more
  computer-intensive computations for the inactive
  predictor groups is implemented on the C++ level via
  OpenMP (\url{http://openmp.org/}).
}
\author{
  Andreas Alfons
}
\seealso{
  \code{\link{coef.seqModel}},
  \code{\link{fitted.seqModel}},
  \code{\link{residuals.seqModel}},
  \code{\link{predict.seqModel}},
  \code{\link{plot.seqModel}}, \code{\link{grplars}}
}
\keyword{regression}

