% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustHD-deprecated.R
\name{robustHD-deprecated}
\alias{robustHD-deprecated}
\alias{sparseLTSGrid}
\alias{sparseLTSGrid.default}
\alias{sparseLTSGrid.formula}
\title{Deprecated functions in package \pkg{robustHD}}
\usage{
sparseLTSGrid(x, ...)

\method{sparseLTSGrid}{formula}(formula, data, ...)

\method{sparseLTSGrid}{default}(x, y, lambda, mode = c("lambda", "fraction"),
  ...)
}
\arguments{
\item{x}{a numeric matrix containing the predictor variables.}

\item{formula}{a formula describing the model.}

\item{data}{an optional data frame, list or environment (or object coercible
to a data frame by \code{\link{as.data.frame}}) containing the variables in
the model.  If not found in data, the variables are taken from
\code{environment(formula)}, typically the environment from which
\code{sparseLTSGrid} is called.}

\item{y}{a numeric vector containing the response variable.}

\item{lambda}{a numeric vector of non-negative values to be used as penalty
parameter.}

\item{mode}{a character string specifying the type of penalty parameter.  If
\code{"lambda"}, \code{lambda} gives the grid of values for the penalty
parameter directly.  If \code{"fraction"}, the smallest value of the penalty
parameter that sets all coefficients to 0 is first estimated based on
bivariate winsorization, then \code{lambda} gives the fractions of that
estimate to be used (hence all values of \code{lambda} should be in the
interval [0,1] in that case).}

\item{\dots}{additional arguments to be passed down, eventually to
\code{\link{sparseLTS}}.}
}
\description{
These functions are provided for compatibility with older versions only, and
may be defunct as soon as the next release.
}
\details{
\code{sparseLTSGrid} is a wrapper function for \code{\link{sparseLTS}}
that only differs in the default values for the penalty parameter
\code{lambda}.
}
\author{
Andreas Alfons
}
\seealso{
\code{\link[base]{Deprecated}}
}
\keyword{regression}
\keyword{robust}

