\name{robustHD-package}
\alias{robustHD-package}
\alias{robustHD}
\docType{package}
\title{
\packageTitle{robustHD}
}
\description{
\packageDescription{robustHD}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{robustHD}
\packageIndices{robustHD}
}
\references{
Alfons (2021) \pkg{robustHD}: An \R package for robust regression with
high-dimensional data. \emph{Journal of Open Source Software}, \bold{6}(67),
3786. \doi{10.21105/joss.03786}.
}
\author{
\packageAuthor{robustHD}

Maintainer: \packageMaintainer{robustHD}
}
\keyword{package}
