\name{robustHD-package}
\alias{robustHD-package}
\alias{robustHD}
\docType{package}
\title{
\packageTitle{robustHD}
}
\description{
\packageDescription{robustHD}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{robustHD}
\packageIndices{robustHD}
}
\author{
\packageAuthor{robustHD}

Maintainer: \packageMaintainer{robustHD}
}
\keyword{package}
