\name{robustHD-package}
\alias{robustHD-package}
\alias{robustHD}
\docType{package}
\title{
Robust methods for high-dimensional data
}
\description{
Robust methods for high-dimensional data, in particular linear
    model selection techniques based on least angle regression and sparse
    regression.
}
\details{
\tabular{ll}{
Package: \tab robustHD\cr
Type: \tab Package\cr
Version: \tab 0.3.0\cr
Date: \tab 2012-02-28\cr
Depends: \tab R (>= 2.14.1), Rcpp (>= 0.9.10), ggplot2 (>= 0.9.2), MASS,
parallel, perry, robustbase (>= 0.6-0)\cr
Imports: \tab RcppArmadillo (>= 0.3.0), pcaPP, stats\cr
LinkingTo: \tab Rcpp, RcppArmadillo\cr
Suggests: \tab mvtnorm\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Andreas Alfons [aut, cre]

Maintainer: Andreas Alfons <alfons@ese.eur.nl>
}
\keyword{package}
